Fix display issues
Now the display value is updated correctly after an operator has been
selected and the user begins to type a new number.

Added the functionalities behind all the numbers and operators.

Cleared up some variable names and rearranged the declarations.

Also rewrote the basic functions so that they are less complicated.